+++ core/src/math/mod.rs
@@ -11,7 +11,7 @@ pub trait Arithmetic:
 	Add<Output = Self>
 	+ Sub<Output = Self>
 	+ Mul<Output = Self>
-	+ Mul<f64, Output = Self>
+	+ Mul<f32, Output = Self>
 	+ Copy
 where
 	Self: Sized,
@@ -22,7 +22,7 @@ impl<T> Arithmetic for T where
 	T: Add<Output = T>
 		+ Sub<Output = T>
 		+ Mul<Output = T>
-		+ Mul<f64, Output = T>
+		+ Mul<f32, Output = T>
 		+ Copy
 		+ Sized
 {
